Kysymyslistausta paranneltu. Refaktoroitu koodia. Koitettu saada maintainabilityä paremmaks
