+++ src/app/components/common/EpicTable/EnhancedTableHead.js
@@ -1,0 +1,56 @@
+import React, { Component } from 'react'
+import TableCell from '@material-ui/core/TableCell'
+import TableHead from '@material-ui/core/TableHead'
+import TableRow from '@material-ui/core/TableRow'
+import TableSortLabel from '@material-ui/core/TableSortLabel'
+import Checkbox from '@material-ui/core/Checkbox'
+import Tooltip from '@material-ui/core/Tooltip'
+import '../../Admin/admin.css'
+
+class EnhancedTableHead extends Component {
+
+  createSortHandler = property => event => {
+    this.props.onRequestSort(event, property)
+  }
+  render () {
+    const { onSelectAllClick, numSelected, order, orderBy, rowCount, rows } = this.props
+    return (
+      <TableHead>
+        <TableRow>
+          <TableCell padding='checkbox'>
+            <Checkbox
+              indeterminate={numSelected > 0 && numSelected < rowCount}
+              checked={numSelected === rowCount}
+              onChange={onSelectAllClick}
+            />
+          </TableCell>
+          {rows.map(row => {
+            return (
+              <TableCell
+                key={row.id}
+                padding={row.disablePadding ? 'none' : 'default'}
+                sortDirection={orderBy === row.id ? order : false}
+              >
+                <Tooltip
+                  title='Sort'
+                  placement={row.numeric ? 'bottom-end' : 'bottom-start'}
+                  enterDelay={300}
+                >
+                  <TableSortLabel
+                    active={orderBy === row.id}
+                    direction={order}
+                    onClick={this.createSortHandler(row.id)}
+                  >
+                    {row.label}
+                  </TableSortLabel>
+                </Tooltip>
+              </TableCell>
+            )
+          })}
+        </TableRow>
+      </TableHead>
+    )
+  }
+}
+
+export default EnhancedTableHead
