+++ src/app/services/courseService.js
@@ -9,27 +9,31 @@ if (process.env.NODE_ENV === 'production') {
 
 const apiUrl = '/api/v1/courses'
 
-const getCourses = async () => {
-  const config = {
-    headers: { 'Authorization': token}
+const setToken = (newToken) => {
+  token = `bearer ${newToken}`
+}
+
+const getConfig = () => {
+  return {
+    headers: { 'Authorization': token }
     // USED FOR DEV PURPOSES
     // params: { force: true }
   }
-  const response = await axios.get(`${baseUrl}${apiUrl}`, config)
+}
+
+const getCourses = async () => {
+  console.log(getConfig())
+  const response = await axios.get(`${baseUrl}${apiUrl}`, getConfig())
   return response.data
 }
 
 const getCourse = async (name) => {
-  const config = {
-    headers: { 'Authorization': token}
-    // USED FOR DEV PURPOSES
-    // params: { force: true }
-  }
-  const response = await axios.get(`${baseUrl}${apiUrl}/${name}`, config)
+  const response = await axios.get(`${baseUrl}${apiUrl}/${name}`, getConfig())
   return response.data
 }
 
 export default {
   getCourses,
-  getCourse
+  getCourse,
+  setToken
 }
